Move PollContentPresenter to io.element.android.features.messages.impl.timeline.factories.event.poll
