+++ features/messages/impl/src/main/kotlin/io/element/android/features/messages/impl/timeline/components/event/TimelineItemPollView.kt
@@ -22,8 +22,8 @@ import androidx.compose.ui.tooling.preview.PreviewParameter
 import io.element.android.features.messages.impl.timeline.model.event.TimelineItemPollContent
 import io.element.android.features.messages.impl.timeline.model.event.TimelineItemPollContentProvider
 import io.element.android.features.poll.api.PollContentView
-import io.element.android.features.poll.api.content.PollContentEvents
-import io.element.android.features.poll.api.content.PollContentState
+import io.element.android.features.messages.impl.timeline.factories.event.poll.PollContentEvents
+import io.element.android.features.messages.impl.timeline.factories.event.poll.PollContentState
 import io.element.android.libraries.designsystem.preview.ElementPreview
 import io.element.android.libraries.designsystem.preview.PreviewsDayNight
 import io.element.android.libraries.matrix.api.core.EventId
