+++ features/messages/impl/src/main/kotlin/io/element/android/features/messages/impl/timeline/factories/event/poll/PollContentState.kt
@@ -14,15 +14,13 @@
  * limitations under the License.
  */
 
-package io.element.android.features.poll.api.content
+package io.element.android.features.messages.impl.timeline.factories.event.poll
 
-import io.element.android.libraries.matrix.api.core.EventId
+import io.element.android.libraries.matrix.api.timeline.item.event.PollContent
 
-sealed interface PollContentEvents {
-    data class OnPollAnswerSelected(
-        val pollStartId: EventId,
-        val answerId: String
-    ) : PollContentEvents
-
-    data object OnPollEndClicked : PollContentEvents
-}
+data class PollContentState(
+    val content: PollContent,
+    val someRandomString: String,
+    val someState: Boolean,
+    val eventSink: (event: PollContentEvents) -> Unit = {},
+)
